Update README.md
Upload a .doc or .docx attachment that includes:

A submission heading:
Your First and Last Name
Professor's Name
Course Name
Date
Assignment Name
Your own answers to the following questions. Use your own words!
What is the difference between the ps command and the top command? Why would you use one over the other?
In the Unix/Linux context, what is a daemon?
Describe the output of the top command to someone who is using it for the first time. Be sure to explain the system summary information and the process table columns.
What is the difference between pressing CTRL+Z and CTRL+C when a program is running in the shell?
Briefly explain the concept of multitasking. In what cases is it an illusion vs. when is it real?